pyproject.toml
[tool.poetry]
name = "home_finder"
version = "0.1.0"
description = "Buy and Sell Properties"
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
uvicorn = { version = "^0.17.0", extras = ["standard"] }
pydantic = {version = "^1.9.0", extras = ["dotenv"]}
yarl = "^1.7.2"
ujson = "^5.1.0"
tortoise-orm = "^0.19.0"
aiomysql = "^0.0.21"
mysqlclient = "^2.0.3"
cryptography = "^36.0.2"
httptools = "^0.3.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^4.0.1"
mypy = "^0.961"
isort = "^5.10.1"
yesqa = "^1.3.0"
pre-commit = "^2.19.0"
wemake-python-styleguide = "^0.16.1"
black = "^22.3.0"
autoflake = "^1.4"
pytest-cov = "^3.0.0"
anyio = "^3.6.1"
pytest-env = "^0.6.2"
asynctest = "^0.13.0"
nest-asyncio = "^1.5.5"
httpx = "^0.22.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["home_finder",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "HOME_FINDER_DB_BASE=home_finder_test",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
